fix: allow new required field with new optional parent and existing grandparent
That case is actually valid. If we have an existing grandparent, and we
add an optional new parent with a new required field, we cannot break
existing users because a) the parent is optional, so the required field
does not mean users have to set it and b) the parent is new so there can
not be previous versions of it that did not set the new required field.